Do not propagate scan level from DNSSPFMechanismHostname
Not propagating the scan level from DNSSPFMechanismHostname to Hostname
will prevent a lot of not directly related objects from being added. It
it also similar to what we already do in other places such as MX records
where we also don't propagate the scanlevel to the target name of the MX
record.

Fixes #673