Enable and disable bits
This adds the variable default_enabled to the BitDefinition which can be
used to disable a bit by default. Two settings BITS_ENABLED and
BITS_DISABLED are added to enable or disable bits.

A cache is added to the get_bit_definition function so we don't import
look up the bits and import them again every time an OOI is updated or a
bit is run.

We also log a warning if there is a directory in the bits directory that
isn't a bit.