Fix human-readable name for ImageMetadata
The Webpage Capture boefje creates images from HostnameHTTPURL OOIs,
which means that after the normalizer creates an ImageMetadata OOI
its reference points to a HostnameHTTPURL OOI instead of an HTTPResource
OOI. minvws/nl-kat-coordination#1560 was created to fix this, but for
now we can fix the human-readable name by trying to the reference as a
HTTPResource first and HostnameHTTPURL if this fails.

This fixes the "IndexError at /en/mispoes/objects/, pop from empty list"
error.