Manage boefjes requirements with Poetry
Move all dependencies from boefjes from `requirements.txt` to
`pyproject.toml` so that they can be managed with Poetry. This means
that these dependencies are now also pinned in the Poetry lock file and
that there are no conflicting dependency versions.

This also pins an older urllib3 version for the `forcediphttpsadapter`
dependency of the `kat_webpage_analysis` boefje.