Fix `poetry-dependencies` target in Makefile
macOS does not like newlines in the for loop, and needs `sed` to have an
(empty) extension for the backup file for the `-i` option.

Additionally, I think we can run `poetry lock --no-update` instead of
`poetry lock --check` to update the Poetry lockfile if required, but
without updating dependencies.