Update normalize.py (kat_snyk_finding_types) for correct RiskLevelSeverity
The severity level is not displayed correctly at this time? Currently, CVEs with a score of 7.x appears as a Medium, however, they should be given a High. Looking at the code, it is because the maximum score is now given in the table instead of the minimum score. By adjusting this, the score would be displayed correctly.